rename some classes from service to mailstore
